+++ Tests/ParseSwiftTests/ParseLiveQueryTests.swift
@@ -544,13 +544,14 @@ class ParseLiveQueryTests: XCTestCase {
         let expectation1 = XCTestExpectation(description: "Send Ping")
         client.sendPing { error in
             XCTAssertEqual(client.isSocketEstablished, false)
-            guard let parseError = error as? ParseError else {
+            guard let urlError = error as? URLError else {
                 XCTFail("Should have casted to ParseError.")
                 expectation1.fulfill()
                 return
             }
-            XCTAssertEqual(parseError.code, ParseError.Code.unknownError)
-            XCTAssertTrue(parseError.message.contains("socket status"))
+            // "Could not connect to the server"
+            // because webSocket connections are not intercepted.
+            XCTAssertEqual(urlError.errorCode, -1004)
             expectation1.fulfill()
         }
         wait(for: [expectation1], timeout: 20.0)